add support for a new nordic special chars found in gbatemp rel
these chars are correct UTF-8 in the XML, but not in the released
usrcheat.dat. apparently they use a similar tool for xml->dat
conversion, which has a set of encoding bugs in each direction.
while at it, make the matching of sharp-s more explicit, as it
occurs only in a single form in the .dat file.