add "tofatxml" mode, to emit no-op lines like in gbatemp rel
using this mode makes it easier to diff against the released xml.
unfortunately, the tool they use inserts additional <date> tags,
and it seems to update the timestamp midway while (as i suspect
that's the workflow) generating the xml from the usrcheat.dat.
the released xml therefore has 2 different timestamps, with the
second half having a timestamp 7 seconds later than the first.
mimicking this behaviour would be very difficult and would
require additional command line options to provide both timestamps
and logic to insert the second timestamp only from a certain
point on.